backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.0
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.1
flask-mail==0.9.1 
twilio==9.2.3
razorpay==1.4.2
redis==5.0.1
openai==1.59.0